[inference] Keep vLLM overflow policy local
Reject an over-limit selected vLLM batch in the vLLM telemetry path, where
whole-batch admission is required to avoid order-dependent dashboard data.
Generic metric publication retains its first-N policy. Collector health reports
the dropped batch while inference continues, and later valid scrapes recover.
